fix(stop-screen-record): prevent redundant notifications
Ensure 'notify-send' is only called if 'wf-recorder' is running, avoiding
unnecessary notifications when the process is not active.

+++ home/.local/bin/stop-screen-record.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
-pkill -f wf-recorder
-
-notify-send 'Screen Recording Stopped'
+if pgrep -f wf-recorder > /dev/null; then
+    pkill -f wf-recorder
+    notify-send 'Screen Recording Stopped'
+fi
